+++ liquidfun/Box2D/Box2D/Dynamics/b2WorldCallbacks.h
@@ -32,6 +32,8 @@ class b2ParticleSystem;
 struct b2ContactResult;
 struct b2Manifold;
 class b2ParticleGroup;
+struct b2ParticleBodyContact;
+struct b2ParticleContact;
 
 /// Joints and fixtures are destroyed when their associated
 /// body is destroyed. Implement this listener so that you
@@ -76,6 +78,31 @@ public:
 	/// Return true if contact calculations should be performed between these two shapes.
 	/// @warning for performance reasons this is only called when the AABBs begin to overlap.
 	virtual bool ShouldCollide(b2Fixture* fixtureA, b2Fixture* fixtureB);
+
+	/// Return true if contact calculations should be performed between a
+	/// fixture and particle.  This is only called if the
+	/// b2_fixtureContactListenerParticle flag is set on the particle.
+	virtual bool ShouldCollide(b2Fixture* fixture,
+							   b2ParticleSystem* particleSystem,
+							   int32 particleIndex)
+	{
+		B2_NOT_USED(fixture);
+		B2_NOT_USED(particleIndex);
+		B2_NOT_USED(particleSystem);
+		return true;
+	}
+
+	/// Return true if contact calculations should be performed between two
+	/// particles.  This is only called if the
+	/// b2_particleContactListenerParticle flag is set on the particle.
+	virtual bool ShouldCollide(b2ParticleSystem* particleSystem,
+							   int32 particleIndexA, int32 particleIndexB)
+	{
+		B2_NOT_USED(particleSystem);
+		B2_NOT_USED(particleIndexA);
+		B2_NOT_USED(particleIndexB);
+		return true;
+	}
 };
 
 /// Contact impulses for reporting. Impulses are used instead of forces because
@@ -107,6 +134,44 @@ public:
 
 	/// Called when two fixtures cease to touch.
 	virtual void EndContact(b2Contact* contact) { B2_NOT_USED(contact); }
+
+	/// Called when a fixture and particle start touching if the
+	/// b2_fixtureContactFilterParticle flag is set on the particle.
+	virtual void BeginContact(b2ParticleSystem* particleSystem,
+							  b2ParticleBodyContact* particleBodyContact)
+	{
+		B2_NOT_USED(particleSystem);
+		B2_NOT_USED(particleBodyContact);
+	}
+
+	/// Called when a fixture and particle stop touching if the
+	/// b2_fixtureContactFilterParticle flag is set on the particle.
+	virtual void EndContact(b2Fixture* fixture,
+							b2ParticleSystem* particleSystem, int32 index)
+	{
+		B2_NOT_USED(fixture);
+		B2_NOT_USED(particleSystem);
+		B2_NOT_USED(index);
+	}
+
+	/// Called when two particles start touching if
+	/// b2_particleContactFilterParticle flag is set on either particle.
+	virtual void BeginContact(b2ParticleSystem* particleSystem,
+							  b2ParticleContact* particleContact)
+	{
+		B2_NOT_USED(particleSystem);
+		B2_NOT_USED(particleContact);
+	}
+
+	/// Called when two particles start touching if
+	/// b2_particleContactFilterParticle flag is set on either particle.
+	virtual void EndContact(b2ParticleSystem* particleSystem,
+							int32 indexA, int32 indexB)
+	{
+		B2_NOT_USED(particleSystem);
+		B2_NOT_USED(indexA);
+		B2_NOT_USED(indexB);
+	}
 
 	/// This is called after a contact is updated. This allows you to inspect a
 	/// contact before it goes to the solver. If you are careful, you can modify the
@@ -160,8 +225,8 @@ public:
 
 	/// Cull an entire particle system from b2World::QueryAABB. Ignored for
 	/// b2ParticleSystem::QueryAABB.
-	/// @return true if you want to include particleSystem in the AABB query, or
-	/// false to cull particleSystem from the AABB query.
+	/// @return true if you want to include particleSystem in the AABB query,
+	/// or false to cull particleSystem from the AABB query.
 	virtual bool ShouldQueryParticleSystem(
 		const b2ParticleSystem* particleSystem)
 	{
@@ -206,9 +271,9 @@ public:
 	///   b2World::RayCast.
 	/// @return <=0 to ignore rest of particle system, fraction to ignore
 	/// particles that are farther away.
-	virtual float32 ReportParticle(	const b2ParticleSystem* particleSystem,
-									int32 index, const b2Vec2& point,
-									const b2Vec2& normal, float32 fraction)
+	virtual float32 ReportParticle(const b2ParticleSystem* particleSystem,
+								   int32 index, const b2Vec2& point,
+								   const b2Vec2& normal, float32 fraction)
 	{
 		B2_NOT_USED(particleSystem);
 		B2_NOT_USED(index);