+++ liquidfun/Box2D/Box2D/Particle/b2ParticleSystem.h
@@ -35,6 +35,8 @@ class b2StackAllocator;
 class b2QueryCallback;
 class b2RayCastCallback;
 class b2Fixture;
+class b2ContactFilter;
+class b2ContactListener;
 struct b2ParticleGroupDef;
 struct b2Vec2;
 struct b2AABB;
@@ -137,8 +139,8 @@ struct b2ParticleSystemDef
 
 	/// Produces additional pressure on repulsive particles
 	/// Larger values repulse more
-	/// Negative values mean attraction. The range where particles behave stably
-	/// is about -0.2 to 2.0.
+	/// Negative values mean attraction. The range where particles behave
+	/// stably is about -0.2 to 2.0.
 	float32 repulsiveStrength;
 
 	/// Produces repulsion between powder particles
@@ -286,13 +288,14 @@ public:
 	/// oldest particles in the system.
 	void SetMaxParticleCount(int32 count);
 
-	/// Pause or unpause the particle system. When paused, b2World::Step() skips
-	/// over this particle system. All b2ParticleSystem function calls still
-	/// work.
+	/// Pause or unpause the particle system. When paused, b2World::Step()
+	/// skips over this particle system. All b2ParticleSystem function calls
+	/// still work.
 	/// @param paused is true to pause, false to un-pause.
 	void SetPaused(bool paused);
 
-	/// @return true if the particle system is being updated in b2World::Step().
+	/// @return true if the particle system is being updated in
+	/// b2World::Step().
 	/// Initially, true, then, the last value passed into SetPaused().
 	bool GetPaused() const;
 
@@ -482,7 +485,8 @@ public:
 	/// Apply an impulse to one particle. This immediately modifies the
 	/// velocity. Similar to b2Body::ApplyLinearImpulse.
 	/// @param index the particle that will be modified.
-	/// @param impulse the world impulse vector, usually in N-seconds or kg-m/s.
+	/// @param impulse the world impulse vector, usually in N-seconds or
+    ///        kg-m/s.
 	void ParticleApplyLinearImpulse(int32 index, const b2Vec2& impulse);
 
 	/// Apply an impulse to all particles between 'firstIndex' and 'lastIndex'.
@@ -493,7 +497,8 @@ public:
 	/// ApplyLinearImpulse(0, 1, impulse).
 	/// @param firstIndex the first particle to be modified.
 	/// @param lastIndex the last particle to be modified.
-	/// @param impulse the world impulse vector, usually in N-seconds or kg-m/s.
+	/// @param impulse the world impulse vector, usually in N-seconds or
+    ///        kg-m/s.
 	void ApplyLinearImpulse(int32 firstIndex, int32 lastIndex,
 							const b2Vec2& impulse);
 
@@ -541,9 +546,12 @@ public:
 	void RayCast(b2RayCastCallback* callback, const b2Vec2& point1,
 				 const b2Vec2& point2) const;
 
+	/// Compute the axis-aligned bounding box for all particles contained
+	/// within this particle system.
+	/// @param aabb Returns the axis-aligned bounding box of the system.
+	void ComputeAABB(b2AABB* const aabb) const;
 
 private:
-
 	friend class b2World;
 	friend class b2ParticleGroup;
 	friend class b2ParticleBodyContactRemovePredicate;
@@ -630,6 +638,175 @@ private:
 		}
 	};
 
+	// *Very* lightweight pair implementation.
+	template<typename A, typename B>
+	struct LightweightPair
+	{
+		A first;
+		B second;
+
+		// Compares the value of two FixtureParticle objects returning
+		// true if left is a smaller value than right.
+		static bool Compare(const LightweightPair& left,
+							const LightweightPair& right)
+		{
+			return left.first < right.first &&
+				left.second < right.second;
+		}
+
+	};
+
+	// Allocator for a fixed set of items.
+	class FixedSetAllocator
+	{
+	public:
+		// Associate a memory allocator with this object.
+		FixedSetAllocator(b2StackAllocator* allocator);
+		// Deallocate storage for this class.
+		~FixedSetAllocator()
+		{
+			Clear();
+		}
+
+		// Allocate internal storage for this object returning the size.
+		int32 Allocate(const int32 itemSize, const int32 count);
+
+		// Deallocate the internal buffer if it's allocated.
+		void Clear();
+
+		// Get the number of items in the set.
+		int32 GetCount() const { return m_count; }
+
+		// Invalidate an item from the set by index.
+		void Invalidate(const int32 itemIndex)
+		{
+			b2Assert(m_valid);
+			m_valid[itemIndex] = 0;
+		}
+
+		// Get the buffer which indicates whether items are valid in the set.
+		const int8* GetValidBuffer() const { return m_valid; }
+
+	protected:
+		// Get the internal buffer.
+		void* GetBuffer() const { return m_buffer; }
+		void* GetBuffer() { return m_buffer; }
+
+		// Reduce the number of items in the set.
+		void SetCount(int32 count)
+		{
+			b2Assert(count <= m_count);
+			m_count = count;
+		}
+
+	private:
+		// Set buffer.
+		void* m_buffer;
+		// Array of size m_count which indicates whether an item is in the
+		// corresponding index of m_set (1) or the item is invalid (0).
+		int8* m_valid;
+		// Number of items in m_set.
+		int32 m_count;
+		// Allocator used to allocate / free the set.
+		b2StackAllocator* m_allocator;
+	};
+
+	// Allocator for a fixed set of objects.
+	template<typename T>
+	class TypedFixedSetAllocator : public FixedSetAllocator
+	{
+	public:
+		// Initialize members of this class.
+		TypedFixedSetAllocator(b2StackAllocator* allocator) :
+			FixedSetAllocator(allocator) { }
+
+		// Allocate a set of objects, returning the new size of the set.
+		int32 Allocate(const int32 numberOfObjects)
+		{
+			Clear();
+			return FixedSetAllocator::Allocate(sizeof(T), numberOfObjects);
+		}
+
+		// Get the index of an item in the set if it's valid return an index
+		// >= 0, -1 otherwise.
+		int32 GetIndex(const T* item) const
+		{
+			if (item)
+			{
+				b2Assert(item >= GetBuffer() &&
+						 item < GetBuffer() + GetCount());
+				const int32 index =
+					(int32)(((uint8*)item - (uint8*)GetBuffer()) /
+							sizeof(*item));
+				if (GetValidBuffer()[index])
+				{
+					return index;
+				}
+			}
+			return -1;
+		}
+
+		// Get the internal buffer.
+		const T* GetBuffer() const
+		{
+			return (const T*)FixedSetAllocator::GetBuffer();
+		}
+		T* GetBuffer() { return (T*)FixedSetAllocator::GetBuffer(); }
+	};
+
+	template<typename T>
+	friend int32 FindItemIndexInFixedSet(
+		const b2ParticleSystem::TypedFixedSetAllocator<T>& set, const T& item);
+
+	// Associates a fixture with a particle index.
+	typedef LightweightPair<b2Fixture*,int32> FixtureParticle;
+
+	// Set of fixture / particle indices.
+	class FixtureParticleSet :
+		public TypedFixedSetAllocator<FixtureParticle>
+	{
+	public:
+		// Initialize members of this class.
+		FixtureParticleSet(b2StackAllocator* allocator) :
+			TypedFixedSetAllocator<FixtureParticle>(allocator) { }
+
+
+		// Initialize from a set of particle / body contacts for particles
+		// that have the b2_fixtureContactListenerParticle flag set.
+		void Initialize(const b2ParticleBodyContact * const bodyContacts,
+						const int32 numBodyContacts,
+						const uint32 * const particleFlagsBuffer);
+
+		// Find the index of a particle / fixture pair in the set or -1
+		// if it's not present.
+		// NOTE: This was not written as a template function to avoid
+		// exposing any dependencies via this header.
+		int32 Find(const FixtureParticle& fixtureParticle) const;
+	};
+
+	// Associates a fixture with a particle index.
+	typedef LightweightPair<int32,int32> ParticlePair;
+
+	// Set of particle / particle pairs.
+	class ParticlePairSet : public TypedFixedSetAllocator<ParticlePair>
+	{
+	public:
+		// Initialize members of this class.
+		ParticlePairSet(b2StackAllocator* allocator) :
+			TypedFixedSetAllocator<ParticlePair>(allocator) { }
+
+		// Initialize from a set of particle contacts.
+		void Initialize(const b2ParticleContact * const contacts,
+						const int32 numContacts,
+						const uint32 * const particleFlagsBuffer);
+
+		// Find the index of a particle pair in the set or -1
+		// if it's not present.
+		// NOTE: This was not written as a template function to avoid
+		// exposing any dependencies via this header.
+		int32 Find(const ParticlePair& pair) const;
+	};
+
 	/// All particle types that require creating pairs
 	static const int32 k_pairFlags =
 		b2_springParticle |
@@ -696,7 +873,9 @@ private:
 
 	void UpdateAllParticleFlags();
 	void UpdateAllGroupFlags();
-	void AddContact(int32 a, int32 b);
+	void AddContact(int32 a, int32 b, b2ContactFilter* const contactFilter,
+					b2ContactListener* const contactListener,
+					ParticlePairSet* const particlePairSet);
 	void UpdateContacts(bool exceptZombie);
 	void UpdateBodyContacts();
 
@@ -733,6 +912,22 @@ private:
 	float32 GetParticleStride() const;
 	float32 GetParticleMass() const;
 	float32 GetParticleInvMass() const;
+
+	// Get the world's contact filter if any particles with the
+	// b2_contactFilterParticle flag are present in the system.
+	b2ContactFilter* GetFixtureContactFilter() const;
+
+	// Get the world's contact filter if any particles with the
+	// b2_particleContactFilterParticle flag are present in the system.
+	b2ContactFilter* GetParticleContactFilter() const;
+
+	// Get the world's contact listener if any particles with the
+	// b2_fixtureContactListenerParticle flag are present in the system.
+	b2ContactListener* GetFixtureContactListener() const;
+
+	// Get the world's contact listener if any particles with the
+	// b2_particleContactListenerParticle flag are present in the system.
+	b2ContactListener* GetParticleContactListener() const;
 
 	template <typename T> void SetUserOverridableBuffer(
 		UserOverridableBuffer<T>* buffer, T* newBufferData, int32 newCapacity);